ver 3.0.5
-UI optimizations
-Automatic configs download and update with main scripts

+++ Elements/Workshop.py
@@ -5,6 +5,9 @@ import tkinter
 import customtkinter
 import requests as requests
 
+from Elements.DownloadUpdate import DownloadWindow
+from Elements.NoUpdatesRequired import NoUpdatesRequired
+
 
 class Workshop:
     def __init__(self, app):
@@ -27,32 +30,34 @@ class Workshop:
         self.image = tkinter.PhotoImage(file="assets\\loading.gif", format=f"gif -index {0}")
         self.ws_load_button = customtkinter.CTkButton(self.workshop_frame_1, text="Download Selected Items",
                                                       command=self.download_workshop_event)
-        self.ws_load_button.grid(row=1, column=0, padx=(20, 10), pady=(10, 0), sticky="nsew")
+        self.ws_load_button.grid(row=1, column=0, padx=(20, 10), pady=(10, 10), sticky="nsew")
 
         self.ws_load_button_processing = customtkinter.CTkButton(self.workshop_frame_1, state="disabled",
                                                                  text="Processing...",
                                                                  image=self.image)
 
-        self.ws_load_cfg_button = customtkinter.CTkButton(self.workshop_frame_1, text="Download Selected Items Сonfigs",
-                                                          command=self.download_cfg_event)
-        self.ws_load_cfg_button.grid(row=2, column=0, padx=(20, 10), pady=(10, 10), sticky="nsew")
-
-        self.ws_load_cfg_button_processing = customtkinter.CTkButton(self.workshop_frame_1, state="disabled",
-                                                                     text="Processing...",
-                                                                     image=self.image)
-
-        self.ws_explorer_button = customtkinter.CTkButton(self.workshop_frame_2, text="Open Folder",
-                                                          command=self.open_folder)
-        self.ws_explorer_button.grid(row=0, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
-
         self.ws_refresh_button = customtkinter.CTkButton(self.workshop_frame_2, text="Refresh List",
                                                          command=self.refresh_event)
-        self.ws_refresh_button.grid(row=1, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
+        self.ws_refresh_button.grid(row=0, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
 
         self.ws_refresh_button_processing = customtkinter.CTkButton(self.workshop_frame_2, state="disabled",
                                                                     text="Processing...",
                                                                     image=self.image)
-        # self.refresh()
+        self.upd_libraries_button = customtkinter.CTkButton(self.workshop_frame_2, text="Update Libraries",
+                                                            command=self.update_libs_event)
+        self.upd_libraries_button.grid(row=1, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
+
+        self.upd_libraries_button_processing = customtkinter.CTkButton(self.workshop_frame_2, state="disabled",
+                                                                       text="Processing...",
+                                                                       image=self.image)
+        self.upd_application_button = customtkinter.CTkButton(self.workshop_frame_2, text="Update Application",
+                                                              command=self.update_application)
+        self.upd_application_button.grid(row=2, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
+
+        self.no_upd_r_window = NoUpdatesRequired()
+        self.no_upd_r_window.protocol("WM_DELETE_WINDOW", self.close_no_upd)
+
+        self.upd_download_window = None
 
     def open_folder(self):
         self.app.check_path()
@@ -71,9 +76,9 @@ class Workshop:
     def download_workshop(self):
         self.condition = False
         self.ws_refresh_button.configure(state="disabled")
-        self.ws_load_cfg_button.configure(state="disabled")
+        self.upd_libraries_button.configure(state="disabled")
         self.ws_load_button.grid_forget()
-        self.ws_load_button_processing.grid(row=1, column=0, padx=(20, 10), pady=(10, 0), sticky="nsew")
+        self.ws_load_button_processing.grid(row=1, column=0, padx=(20, 10), pady=(10, 10), sticky="nsew")
         self.app.after(50, self.processing_animation_load, self.condition)
 
         self.app.check_path()
@@ -85,37 +90,46 @@ class Workshop:
                     f.write(r.content)
                 switch.configure(border_color="#31dea4", fg_color="#31dea4")
 
+        self.download_cfg()
         self.condition = True
         self.ws_refresh_button.configure(state="normal")
-        self.ws_load_cfg_button.configure(state="normal")
+        self.upd_libraries_button.configure(state="normal")
         self.ws_load_button_processing.grid_forget()
-        self.ws_load_button.grid(row=1, column=0, padx=(20, 10), pady=(10, 0), sticky="nsew")
-
-    def download_cfg_event(self):
-        threading._start_new_thread(self.download_cfg, ())
+        self.ws_load_button.grid(row=1, column=0, padx=(20, 10), pady=(10, 10), sticky="nsew")
 
     def download_cfg(self):
-        self.condition = False
-        self.ws_refresh_button.configure(state="disabled")
-        self.ws_load_button.configure(state="disabled")
-        self.ws_load_cfg_button.grid_forget()
-        self.ws_load_cfg_button_processing.grid(row=2, column=0, padx=(20, 10), pady=(10, 10), sticky="nsew")
-        self.app.after(50, self.processing_animation_load_cfg, self.condition)
+        def cfg_equal(script):
+            if not os.path.exists(cfg):
+                return False
+            with open(self.app.settings_path + f"\\cfg_{script}") as file:
+                current_cfg = file.readlines()
+            for i in range(len(current_cfg)):
+                current_cfg[i] = current_cfg[i].replace(" ", "").rstrip()
+                if "=" in current_cfg[i]:
+                    current_cfg[i] = current_cfg[i][:current_cfg[i].find("=")] + current_cfg[i][
+                        current_cfg[i].rfind(";")]
 
-        self.app.check_path()
+            online_cfg = requests.get(f"https://raw.githubusercontent.com/Lazy-World/warframe-ahk/main/settings/"
+                                      f"{'cfg_' + script}").content.decode("utf-8").splitlines()
+
+            for i in range(len(online_cfg)):
+                online_cfg[i] = online_cfg[i].replace(" ", "").rstrip()
+                if "=" in online_cfg[i]:
+                    online_cfg[i] = online_cfg[i][:online_cfg[i].find("=")] + online_cfg[i][online_cfg[i].rfind(";")]
+
+            if online_cfg == current_cfg:
+                return True
+            else:
+                return False
+
         for switch in self.scrollable_frame_switches_2:
             if switch.get() == 1:
                 cfg = self.app.settings_path + "\\cfg_" + switch.cget("text")
-                rq = requests.get(f"https://raw.githubusercontent.com/Lazy-World/warframe-ahk/main/settings/"
-                                  f"{'cfg_'+switch.cget('text')}")
-                with open(cfg, 'wb') as f:
-                    f.write(rq.content)
-
-        self.condition = True
-        self.ws_refresh_button.configure(state="normal")
-        self.ws_load_button.configure(state="normal")
-        self.ws_load_cfg_button_processing.grid_forget()
-        self.ws_load_cfg_button.grid(row=2, column=0, padx=(20, 10), pady=(10, 10), sticky="nsew")
+                if not os.path.exists(cfg) or not cfg_equal(switch.cget('text')):
+                    rq = requests.get(f"https://raw.githubusercontent.com/Lazy-World/warframe-ahk/main/settings/"
+                                      f"{'cfg_' + switch.cget('text')}")
+                    with open(cfg, 'wb') as f:
+                        f.write(rq.content)
 
     def refresh_event(self):
         threading._start_new_thread(self.refresh, ())
@@ -124,13 +138,6 @@ class Workshop:
         if not self.condition:
             self.ws_refresh_button_processing.configure(image=next(self.app.reload_button_icon_anim))
             self.app.after(50, self.processing_animation_refresh, self.condition)
-        else:
-            return
-
-    def processing_animation_load_cfg(self, condition):
-        if not self.condition:
-            self.ws_load_cfg_button_processing.configure(image=next(self.app.reload_button_icon_anim))
-            self.app.after(50, self.processing_animation_load_cfg, self.condition)
         else:
             return
 
@@ -144,9 +151,9 @@ class Workshop:
     def refresh(self):
         self.condition = False
         self.ws_load_button.configure(state="disabled")
-        self.ws_load_cfg_button.configure(state="disabled")
+        self.upd_libraries_button.configure(state="disabled")
         self.ws_refresh_button.grid_forget()
-        self.ws_refresh_button_processing.grid(row=1, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
+        self.ws_refresh_button_processing.grid(row=0, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
         self.app.after(50, self.processing_animation_refresh, self.condition)
 
         if self.scrollable_frame_switches_2:
@@ -181,6 +188,59 @@ class Workshop:
 
         self.condition = True
         self.ws_load_button.configure(state="normal")
-        self.ws_load_cfg_button.configure(state="normal")
+        self.upd_libraries_button.configure(state="normal")
         self.ws_refresh_button_processing.grid_forget()
-        self.ws_refresh_button.grid(row=1, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
+        self.ws_refresh_button.grid(row=0, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
+
+    def update_libs_event(self):
+        threading._start_new_thread(self.update_libs, ())
+
+    def update_libs(self):
+        self.condition = False
+        self.ws_load_button.configure(state="disabled")
+        self.ws_refresh_button.configure(state="disabled")
+        self.upd_libraries_button.grid_forget()
+        self.upd_libraries_button_processing.grid(row=1, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
+        self.app.after(50, self.processing_animation_libs, self.condition)
+
+        self.app.check_path()
+        answer = requests.get("https://api.github.com/repos/Lazy-World/warframe-ahk/contents/libraries").content
+        json_answer = json.loads(answer.decode("utf-8"))
+        for item in json_answer:
+            if item["name"] != "game_settings.ahk" and item["name"] != "key_decode.ahk":
+                filename = self.app.lib_path + "\\" + item["name"]
+                r = requests.get(item["download_url"])
+                with open(filename, 'wb') as f:
+                    f.write(r.content)
+
+        self.condition = True
+        self.ws_refresh_button.configure(state="normal")
+        self.ws_load_button.configure(state="normal")
+        self.upd_libraries_button_processing.grid_forget()
+        self.upd_libraries_button.grid(row=1, column=1, padx=(10, 20), pady=(10, 0), sticky="nsew")
+
+    def processing_animation_libs(self, condition):
+        if not self.condition:
+            self.upd_libraries_button_processing.configure(
+                image=next(self.app.reload_button_icon_anim))
+            self.app.after(50, self.processing_animation_libs, self.condition)
+        else:
+            return
+
+    def update_application(self):
+        if self.app.version == self.app.online_version:
+            self.no_upd_r_window.geometry(f"+{self.app.winfo_rootx() + 500}+{self.app.winfo_rooty()}")
+            self.no_upd_r_window.grab_set()
+            self.no_upd_r_window.deiconify()
+        else:
+            self.upd_download_window = DownloadWindow(self.app)
+            self.upd_download_window.protocol("WM_DELETE_WINDOW", self.close_upd)
+            threading._start_new_thread(self.upd_download_window.download, ())
+
+    def close_no_upd(self):
+        self.no_upd_r_window.grab_release()
+        self.no_upd_r_window.withdraw()
+
+    def close_upd(self):
+        self.upd_download_window.grab_release()
+        self.upd_download_window.destroy()
